fix: common.sh evaluate with overwritten /vars
This allows to overwrite the static values in /vars, without having
to also copy-paste the dynamic variables based on the previous ones.

For instance, `build_dir` is simply a concatenation of 2 variables,
as such, it should not have to be copypasted in /vars, but before
this fix, that was necessary.